Refactor of base
Decided to do a cleanup of the entire dotfiles, this includes getting
rid of eww configs as I do not use it anymore(pain that I put time into
that), rewriting the media script, removed upload functionality for
now(pain), and switched to hyprshot from grimshot. Also modified a bunch
of other things to just be cleaner/more documented.